# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GamePrice API
  description: |
    Real-time game price comparison across 8 digital stores.
    
    Search for any game and get the best prices from Eneba, Instant Gaming, 
    CheapShark, G2A, Kinguin, CDKeys, PSN Store, and Xbox Store.
    
    ## Authentication
    All endpoints require an `X-API-Key` header (except `/health`).
    
    ## Plans
    | Plan | Daily Requests | Rate Limit | Price |
    |------|---------------|------------|-------|
    | Free | 100 | 10/min | $0 |
    | Basic | 10,000 | 60/min | $9.99/mo |
    | Pro | 100,000 | 120/min | $29.99/mo |
    | Ultra | Unlimited | 300/min | $99.99/mo |
  version: "1.0.0"
  contact:
    name: GamePrice Support
    email: support@gameprice.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://baratoya.billytech.es/api/v1
    description: Production server
  - url: https://baratoya.billytech.es/api
    description: Health endpoint

paths:
  /health:
    get:
      tags:
        - System
      summary: Health check
      description: Check if the API is running. No authentication required.
      operationId: healthCheck
      security: []
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /stores:
    get:
      tags:
        - Stores
      summary: List all stores
      description: Returns all supported game stores with their currency information.
      operationId: listStores
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stores:
                    type: array
                    items:
                      $ref: "#/components/schemas/Store"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /search:
    get:
      tags:
        - Search
      summary: Search across all stores
      description: |
        Searches all 8 stores simultaneously and returns the best prices.
        Results are sorted by price (lowest first).
        
        **Cache behavior:**
        - First request for a query: ~7s (live scraping)
        - Cached results: <0.1s (30 min TTL)
      operationId: searchAll
      security:
        - ApiKeyAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Game name to search
          schema:
            type: string
            example: "elden ring"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  query:
                    type: string
                    example: "elden ring"
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/PriceResult"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

  /prices/{store}:
    get:
      tags:
        - Search
      summary: Search by store
      description: Search for a game in a specific store only.
      operationId: searchByStore
      security:
        - ApiKeyAuth: []
      parameters:
        - name: store
          in: path
          required: true
          description: Store ID to search
          schema:
            type: string
            enum:
              - eneba
              - instant-gaming
              - cheapshark
              - g2a
              - kinguin
              - cdkeys
              - psn-store
              - xbox-store
        - name: q
          in: query
          required: true
          description: Game name to search
          schema:
            type: string
            example: "witcher"
      responses:
        "200":
          description: Store-specific results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  query:
                    type: string
                    example: "witcher"
                  store:
                    type: string
                    example: "cheapshark"
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/PriceResult"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"

  /deals:
    get:
      tags:
        - Deals
      summary: Current deals
      description: |
        Returns the best current deals from CheapShark.
        PC games under $15 with the biggest savings.
        Cached for 15 minutes.
      operationId: getDeals
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/DealResult"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Your API key from the RapidAPI dashboard

  schemas:
    Store:
      type: object
      properties:
        id:
          type: string
          example: "eneba"
        name:
          type: string
          example: "Eneba"
        url:
          type: string
          format: uri
          example: "https://www.eneba.com"
        currency:
          type: string
          example: "EUR"

    PriceResult:
      type: object
      properties:
        store:
          type: string
          example: "Eneba"
        name:
          type: string
          example: "ELDEN RING"
        price:
          type: number
          format: float
          example: 29.99
        original_price:
          type: number
          format: float
          example: 59.99
        discount_percent:
          type: integer
          example: 50
        currency:
          type: string
          example: "EUR"
        url:
          type: string
          format: uri
          example: "https://www.eneba.com/..."
        in_stock:
          type: boolean
          example: true
        platform:
          type: string
          example: "Steam"

    DealResult:
      type: object
      properties:
        store:
          type: string
          example: "CheapShark"
        name:
          type: string
          example: "GravNewton"
        price:
          type: number
          format: float
          example: 0.49
        original_price:
          type: number
          format: float
          example: 9.99
        discount_percent:
          type: integer
          example: 95
        currency:
          type: string
          example: "USD"
        url:
          type: string
          format: uri
        platform:
          type: string
          example: "PC"
        steam_rating:
          type: number
          format: float
          example: 85

    Meta:
      type: object
      properties:
        count:
          type: integer
          example: 66
        stores_searched:
          type: integer
          example: 8
        cached:
          type: boolean
          example: true

    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Missing API key. Provide X-API-Key header."

  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "Missing API key. Provide X-API-Key header."

    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "Missing required parameter: q"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "Unknown store. Available stores: eneba, instant-gaming, cheapshark, g2a, kinguin, cdkeys, psn-store, xbox-store"

    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Maximum requests allowed per minute
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
        X-Plan:
          schema:
            type: string
          description: Your current plan tier
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "Rate limit exceeded. Maximum 10 requests per minute for your plan (free)."